+++ Changelog.md
@@ -1,6 +1,12 @@
 # Changelog for CimFolder
 
+## v0.2.0
+
++ Added a new type extension to `cimFile` called `ModifiedAge`.
++ Added a new view called `Age` to use the `ModifiedAge` property.
++ Re-ordered ANSI formatting.
++ Updated `README.md`.
+
 ## v0.1.0
 
 + Initial files
-